Extract service creation to a class
